+++ src/utils.hpp
@@ -84,6 +84,6 @@ uint64_t harmonic_mean(vector<uint64_t>& counts);
 
 string getRealPath(string file, string& dir);
 
-uint get_color_number(string& fof);
+uint16_t get_color_number(string& fof);
 
 #endif
